split provider routers into modules
The router modules in anthropic and openai were doing too much, I have
divided the logic from the rest implementation

also, ive moved the models into their respective modules, and made
common functions work on generic dicts instead of specific models to
avoid typing headaches
