+++ mockai/gui/static/index.html
